removed bugs, fixed error calculation, cleaned up comments

+++ Core/include/ACTS/Seeding/BinFinder.hpp
@@ -15,7 +15,7 @@ public:
 
    virtual
    std::set<size_t>
-   findBins(int phiBin, int zBin, std::unique_ptr<SPGrid>& binnedSP);
+   findBins(size_t phiBin, size_t zBin, std::unique_ptr<SPGrid>& binnedSP);
 
 };
 }
