+++ Core/src/Seeding/BinFinder.cpp
@@ -2,6 +2,6 @@
 
 //std::set<std::array<int, 2> >
 std::set<size_t>
-Acts::Seeding::BinFinder::findBins(int i, int j, std::unique_ptr<Acts::Seeding::SPGrid>& binnedSP){
+Acts::Seeding::BinFinder::findBins(size_t i, size_t j, std::unique_ptr<Acts::Seeding::SPGrid>& binnedSP){
   return binnedSP->neighborHoodIndices({i,j});  
 }
